# crates/lambdas/indexer-api/openapi.yaml
openapi: 3.0.3
info:
  title: Boundless Indexer API
  description: API for accessing staking, delegation, and Proof of Verifiable Work (PoVW) data for the Boundless protocol.
  version: 1.0.0
  contact:
    name: Boundless Development Team
servers:
  - url: /
    description: Current server
...
paths:
...
  /v1/staking:
    get:
      tags:
        - Staking
      summary: Get Staking Summary Statistics
      description: Get global staking summary statistics across all epochs
      operationId: getStakingSummary
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StakingAggregateResponse'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /v1/staking/addresses:
    get:
      tags:
        - Staking
      summary: Get Aggregate Staking Leaderboard
      description: Get the aggregate staking positions leaderboard
      operationId: getStakingLeaderboard
      parameters:
        - $ref: '#/components/parameters/LimitParam'
        - $ref: '#/components/parameters/OffsetParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StakingAggregateResponse'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /v1/staking/epochs:
    get:
      tags:
        - Staking
      summary: Get All Epochs Staking Summary
      description: Get a list of all epochs with staking summary statistics
      operationId: getAllEpochsStakingSummary
      parameters:
        - $ref: '#/components/parameters/LimitParam'
        - $ref: '#/components/parameters/OffsetParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StakingEpochsSummaryResponse'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /v1/staking/epochs/{epoch}:
    get:
      tags:
        - Staking
      summary: Get Epoch Staking Summary
      description: Get staking summary statistics for a specific epoch
      operationId: getStakingByEpoch
      parameters:
        - $ref: '#/components/parameters/EpochParam'
        - $ref: '#/components/parameters/LimitParam'
        - $ref: '#/components/parameters/OffsetParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StakingByEpochResponse'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /v1/staking/epochs/{epoch}/addresses:
    get:
      tags:
        - Staking
      summary: Get Epoch Staking Positions
      description: Get staking positions for a specific epoch
      operationId: getEpochStakingPositions
      parameters:
        - $ref: '#/components/parameters/EpochParam'
        - $ref: '#/components/parameters/LimitParam'
        - $ref: '#/components/parameters/OffsetParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StakingByEpochResponse'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /v1/staking/addresses/{address}:
    get:
      tags:
        - Staking
      summary: Get Staking History by Address
      description: Get complete staking history for a specific address across all epochs
      operationId: getStakingHistoryByAddress
      parameters:
        - $ref: '#/components/parameters/AddressParam'
        - $ref: '#/components/parameters/LimitParam'
        - $ref: '#/components/parameters/OffsetParam'
        - name: start_epoch
          in: query
          description: Filter results from this epoch onwards
          schema:
            type: integer
            minimum: 0
        - name: end_epoch
          in: query
          description: Filter results up to this epoch
          schema:
            type: integer
            minimum: 0
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StakingByEpochResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /v1/staking/epochs/{epoch}/addresses/{address}:
    get:
      tags:
        - Staking
      summary: Get Staking Position by Address and Epoch
      description: Get staking position for a specific address at a specific epoch
      operationId: getStakingByAddressAndEpoch
      parameters:
        - $ref: '#/components/parameters/EpochParam'
        - $ref: '#/components/parameters/AddressParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StakingPositionByEpoch'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  parameters:
    LimitParam:
      name: limit
      in: query
      description: Maximum number of results to return
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 50
    OffsetParam:
      name: offset
      in: query
      description: Number of results to skip for pagination
      schema:
        type: integer
        minimum: 0
        default: 0
    EpochParam:
      name: epoch
      in: path
      required: true
      description: Epoch number (2-day period)
      schema:
        type: integer
        minimum: 0
    AddressParam:
      name: address
      in: path
      required: true
      description: Ethereum address (0x-prefixed, checksummed)
      schema:
        type: string
        pattern: '^0x[a-fA-F0-9]{40}$'
        example: '0x2408e37489c231f883126c87e8aadbad782a040a'
  schemas:
    Pagination:
      type: object
      required:
        - count
        - offset
        - limit
      properties:
        count:
          type: integer
          description: Number of items in this response
        offset:
          type: integer
          description: Offset used for this response
        limit:
          type: integer
          description: Maximum items per response
    StakingPositionAggregate:
      type: object
      required:
        - staker_address
        - total_staked
        - is_withdrawing
        - epochs_participated
      properties:
        rank:
          type: integer
          nullable: true
          description: Leaderboard rank (only present in leaderboard contexts)
        staker_address:
          type: string
          description: Ethereum address of the staker
        total_staked:
          type: string
          description: Total amount staked (in wei)
        is_withdrawing:
          type: boolean
          description: Whether the staker is withdrawing
        rewards_delegated_to:
          type: string
          nullable: true
          description: Address rewards are delegated to
        votes_delegated_to:
          type: string
          nullable: true
          description: Address votes are delegated to
        epochs_participated:
          type: integer
          description: Number of epochs participated
    StakingPositionByEpoch:
      type: object
      required:
        - staker_address
        - epoch
        - staked_amount
        - is_withdrawing
      properties:
        rank:
          type: integer
          nullable: true
          description: Leaderboard rank (only present in leaderboard contexts, not in individual address queries)
        staker_address:
          type: string
          description: Ethereum address of the staker
        epoch:
          type: integer
          description: Epoch number
        staked_amount:
          type: string
          description: Amount staked in this epoch (in wei)
        is_withdrawing:
          type: boolean
          description: Whether the staker is withdrawing
        rewards_delegated_to:
          type: string
          nullable: true
          description: Address rewards are delegated to
        votes_delegated_to:
          type: string
          nullable: true
          description: Address votes are delegated to
...
    StakingAggregateResponse:
      type: object
      required:
        - entries
        - pagination
      properties:
        entries:
          type: array
          items:
            $ref: '#/components/schemas/StakingPositionAggregate'
        pagination:
          $ref: '#/components/schemas/Pagination'
    StakingByEpochResponse:
      type: object
      required:
        - entries
        - pagination
      properties:
        entries:
          type: array
          items:
            $ref: '#/components/schemas/StakingPositionByEpoch'
        pagination:
          $ref: '#/components/schemas/Pagination'
    StakingEpochsSummaryResponse:
      type: object
      required:
        - entries
        - pagination
      properties:
        entries:
          type: array
          items:
            $ref: '#/components/schemas/EpochStakingSummary'
        pagination:
          $ref: '#/components/schemas/Pagination'
    EpochStakingSummary:
      type: object
      required:
        - epoch
        - total_staked
        - total_staked_formatted
        - num_stakers
        - num_withdrawing
        - total_staking_emissions
        - total_staking_emissions_formatted
        - total_staking_power
        - total_staking_power_formatted
        - num_reward_recipients
        - epoch_start_time
        - epoch_end_time
      properties:
        epoch:
          type: integer
          description: Epoch number
        total_staked:
          type: string
          description: Total amount staked in this epoch (in wei)
        total_staked_formatted:
          type: string
          description: Total amount staked (human-readable)
        num_stakers:
          type: integer
          description: Number of stakers in this epoch
        num_withdrawing:
          type: integer
          description: Number of stakers withdrawing
        total_staking_emissions:
          type: string
          description: Total staking emissions for this epoch (in wei)
        total_staking_emissions_formatted:
          type: string
          description: Total staking emissions (human-readable)
        total_staking_power:
          type: string
          description: Total staking power in this epoch (in wei)
        total_staking_power_formatted:
          type: string
          description: Total staking power (human-readable)
        num_reward_recipients:
          type: integer
          description: Number of unique reward recipients
        epoch_start_time:
          type: integer
          description: Unix timestamp when epoch started
        epoch_end_time:
          type: integer
          description: Unix timestamp when epoch ends
        last_updated_at:
          type: string
          nullable: true
          description: Last update timestamp
...
    ErrorResponse:
      type: object
      required:
        - error
        - message
      properties:
        error:
          type: string
          description: Error type
        message:
          type: string
          description: Detailed error message
  responses:
    BadRequest:
      description: Bad Request - Invalid parameters
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: Bad Request
            message: Invalid address format
    NotFound:
      description: Not Found - Resource does not exist
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: Not Found
            message: No data found for the specified parameters
    InternalServerError:
      description: Internal Server Error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: Internal Server Error
            message: An unexpected error occurred